% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{checkTERT}
\alias{checkTERT}
\title{Title: Chek TERT promoter mutations}
\usage{
checkTERT(df)
}
\arguments{
\item{df}{Data frame containing mutations per sample (mutect2/strelka2 etc.)}
}
\description{
This function shows TERT promoter mutations C228T and C250T.
These mutations are common in gliomas and other cancers.
It filters the input data frame for mutations in the TERT promoter region at specific positions
}
