% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{covStats}
\alias{covStats}
\title{Title: Coverage Statistics}
\usage{
covStats(df, charTerm)
}
\arguments{
\item{df}{Data frame containing sample coverage data from metrics (metrics/all.wgs_metrics.txt)}

\item{charTerm}{usually A or B for norm and tumor}
}
\description{
This function calculate coverage statistics for samples. You can filter by a specific trailing character.\if{html}{\out{<br>}}
OSS: expected SAMPLE and MEAN_COVERAGE columns
}
