% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{FGA}
\alias{FGA}
\title{Title: FGA (Fraction of the genome altered)}
\usage{
FGA(df, threshold = 0.2)
}
\arguments{
\item{df}{Data frame containing cncf info from facets (facets/all.cncf.txt)}

\item{threshold}{Value to define a segment altered or not.
It is the absolute value of cnlr.median. (default 0.2)}
}
\description{
This function calculates the fraction of the genome altered (FGA) based on cncf data from facets. \if{html}{\out{<br>}}
OSS: expected TUMOR_NORMAL, start, end and cnlr.median columns
}
