% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{make_mut_landscape}
\alias{make_mut_landscape}
\title{Make Mutation Landscape}
\usage{
make_mut_landscape(df, landmark)
}
\arguments{
\item{df}{Data frame containing mutations with columns TUMOR_SAMPLE, ANN....GENE, and annotation. Filtere removing not interesting mutations.}

\item{landmark}{Character vector of landmark genes to filter the mutations.}
}
\value{
A data frame representing the mutation landscape, where rows are genes and columns are samples
}
\description{
Create a mutation landscape matrix from a dataframe of mutations.Samples are cols, genes are rows.

This function constructs a mutation landscape matrix where each cell contains the combined annotations
for a specific gene in a specific sample. It filters the mutations based on the provided landmark genes,
and combines annotations for each sample and gene. The resulting matrix is cleaned to remove unnecessary
annotations and ensure proper formatting.
}
