% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{purStats}
\alias{purStats}
\title{Title: Purity Statistics}
\usage{
purStats(df, charTerm = "")
}
\arguments{
\item{df}{Data frame containing sample statistics of data from facets (facets/cncf/all.summary.txt)}

\item{charTerm}{usually A or B for norm and tumor or ORG for organoids.}
}
\description{
Calculate purity statistics \if{html}{\out{<br>}}
Not take in account 0.3 and NA values \if{html}{\out{<br>}}
OSS: expected Purity column
}
