% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{rename_anno}
\alias{rename_anno}
\title{Rename Annotation}
\usage{
rename_anno(annotation, snpeff)
}
\arguments{
\item{annotation}{Character string to be renamed. Typically a concatenation of variant effect terms.}

\item{snpeff}{Data frame containing SnpEff effect ontology and ranks. Default is the bundled snpeff.xlsx
from the package's \code{ref} folder. Need to be passed as argument!}
}
\value{
A character string with the renamed annotation.
}
\description{
Rename the annotation string based on SnpEff data.

This function splits the annotation string (potentially multiple annotations joined by '&'),
looks up their ranks in the provided SnpEff dataframe, and chooses the highest-priority
(lowest rank) annotation. It then recodes some specific labels to 'Other' for consistency.
}
