% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{TMB}
\alias{TMB}
\title{Title: Tumor Mutational Burden (TMB)}
\usage{
TMB(df)
}
\arguments{
\item{df}{Data frame containing mutations per sample (mutect2/strelka2 etc.)}
}
\description{
This function calculates the TMB based on the number of mutations per sample.
OSS :It assumes that the input data frame has a column named TUMOR_SAMPLE.
}
