% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsFun.r
\name{WGD}
\alias{WGD}
\title{Title: Whole Genome Douplication (WGD)}
\usage{
WGD(facets, threshold = 2.6)
}
\arguments{
\item{facets}{Data frame containing facets data (facets/all.summary.txt)}

\item{threshold}{Value to define a sample as WGD or diploid (default 2.6)}
}
\description{
This function calculates the whole genome duplication (WGD) status based on ploidy values from facets.
OSS: expected tumor and Ploidy columns
}
